+++ src/main/java/one/util/streamex/AbstractStreamEx.java
@@ -267,6 +267,23 @@ public abstract class AbstractStreamEx<T, S extends AbstractStreamEx<T, S>> exte
         return new DoubleStreamEx(stream().mapToDouble(mapper), context);
     }
 
+    /**
+     * Returns an {@link EntryStream} consisting of the {@link Entry} objects
+     * which keys are elements of this stream and values are results of applying
+     * the given function to the elements of this stream.
+     *
+     * <p>
+     * This is an <a href="package-summary.html#StreamOps">intermediate</a>
+     * operation.
+     *
+     * @param <V> The {@code Entry} value type
+     * @param keyMapper a non-interfering, stateless function to apply to each
+     *        element
+     * @return the new stream
+     */
+    public <K> EntryStream<K, T> mapToEntryByKey(Function<? super T, ? extends K> keyMapper) {
+        return new EntryStream<>(stream().map(e -> new AbstractMap.SimpleImmutableEntry<>(keyMapper.apply(e), e)), context);
+    }
     
     /**
      * Returns an {@link EntryStream} consisting of the {@link Entry} objects
@@ -282,9 +299,10 @@ public abstract class AbstractStreamEx<T, S extends AbstractStreamEx<T, S>> exte
      *        element
      * @return the new stream
      */
-    public <K, V> EntryStream<K, V> mapToEntry(Function<? super T, ? extends Map.Entry<K, V>> mapper) {
-        return new EntryStream<>(stream().map(e -> mapper.apply(e)), context);
+    public <V> EntryStream<T, V> mapToEntryByVal(Function<? super T, ? extends V> valueMapper) {
+        return new EntryStream<>(stream().map(e -> new AbstractMap.SimpleImmutableEntry<>(e, valueMapper.apply(e))), context);
     }
+
     /**
      * Returns an {@link EntryStream} consisting of the {@link Entry} objects
      * which keys and values are results of applying the given functions to the